MoveArmGame version 2.0.1 (adapting Hold game)

+++ src/main/java/com/example/move_arm/database/ClickDao.java
@@ -25,7 +25,7 @@ public class ClickDao {
                 ps.setDouble(5, cd.getCursor().getY());
                 ps.setDouble(6, cd.getCenter().getX());
                 ps.setDouble(7, cd.getCenter().getY());
-                ps.setDouble(8, cd.getRadius());
+                ps.setInt(8, cd.getRadius());
                 ps.addBatch();
             }
             ps.executeBatch();
